reference/requirements.txt
attrs==21.2.0
Automat==20.2.0
awscrt==0.21.1
awsiotsdk==1.22.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
boto3==1.35.7
botocore==1.35.7
Brotli==1.1.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
# cloud-init==23.2.2
colorama==0.4.4
# command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
# dbus-python==1.2.18
distro==1.7.0
# distro-info==1.1+ubuntu0.1
dotenv==0.9.9
# ec2-hibinit-agent==1.0.0
envs==1.4
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
# hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
locust==2.19.0
MarkupSafe==2.1.3
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
paho-mqtt==1.6.1
pandas==2.1.3
pexpect==4.8.0
psutil==5.9.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycognito==2024.5.1
pycparser==2.22
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.9.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
# python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
# python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
pyzmq==25.1.1
requests==2.32.3
roundrobin==0.0.4
s3transfer==0.10.2
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
# sos==4.5.6
ssh-import-id==5.11
# systemd-python==234
Twisted==22.1.0
tzdata==2023.3
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
zope.event==5.0
zope.interface==5.4.0
